changed `commands` into a package and moved the invocation point back into the root `main` package
