+++ commands/package_template_build.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"context"
