+++ commands/package_uninstall.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"github.com/pkg/errors"
