+++ commands/server_init.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	appRuntime "runtime"
